feat: improve performance of colored logs
Using simple ANSI escape sequence codes, this commit replaces the
library used to color the text. This significantly improves the
benchmarks for colored logs keeping the allocs/op exactly same as non
colored ones and minimal increase in ns/op as compared to non colored
ones.

Unintended side effect is that this is now a lib with zero dependencies!